requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-google-genai
langchain-pinecone
pinecone-client
pypdf
beautifulsoup4
python-dotenv
playwright
mangum
boto3
email-validator 
phonenumbers